Backporting a bugfix from version 2.0.2 that make folder_contents crazy
Avoid using a mutable default argument in the FolderContentsTable code.
In a LinguaPlone environment after viewing the folder contents of a
collection, the language of that collection got stuck as a content
filter and wasn't reset anymore.
Viewing the folder contents of any item in a different language showed
an empty table until the Zope instance was restarted.