Address plone/Products.CMFPlone#383
Provide a test that proves that deleting from the folder contents view works properly

Solve the problem for the delete action view by comparing the inner and outer acquisition chains of the object to be deleted.  If they are different, do not delete because the object is not the one we wanted to be rid of.